+++ src/modules/internalServerError.module.ts
@@ -1,7 +1,6 @@
 import {CommonModule} from '@angular/common';
-import {NgModule, ModuleWithProviders} from '@angular/core';
+import {NgModule} from '@angular/core';
 import {InternalServerErrorComponent} from '../internalServerError/internalServerError.component';
-import {InternalServerErrorService} from '../internalServerError/internalServerError.service';
 
 /**
  * Module for internal server error handling
@@ -14,16 +13,4 @@ import {InternalServerErrorService} from '../internalServerError/internalServerE
 })
 export class InternalServerErrorModule
 {
-    //######################### public methods #########################
-    
-    /**
-     * Returns module with internal server error service
-     */
-    public static forRoot(): ModuleWithProviders 
-    {
-        return {
-            ngModule: InternalServerErrorModule,
-            providers: [InternalServerErrorService]
-        };
-    }
 }