+++ src/modules/serverValidations.module.ts
@@ -1,9 +1,8 @@
 import {CommonModule} from '@angular/common';
-import {NgModule, ModuleWithProviders} from '@angular/core';
+import {NgModule} from '@angular/core';
 import {ServerValidationValidatorDirective} from '../serverValidation/serverValidationValidator.directive';
 import {ServerValidationMessagesComponent} from '../serverValidation/serverValidationMessages.component';
 import {WithServerValidationsDirective} from '../serverValidation/withServerValidations.directive';
-import {ServerValidationService} from '../serverValidation/serverValidation.service';
 
 /**
  * Module for server validations
@@ -16,16 +15,4 @@ import {ServerValidationService} from '../serverValidation/serverValidation.serv
 })
 export class ServerValidationsModule
 {
-    //######################### public methods #########################
-    
-    /**
-     * Returns module with server validation service
-     */
-    public static forRoot(): ModuleWithProviders 
-    {
-        return {
-            ngModule: ServerValidationsModule,
-            providers: [ServerValidationService]
-        };
-    }
 }